requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
haversine==2.5.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
Cython==0.29.24
# pandas==0.23.1
pipreqs==0.4.11
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
yarg==0.1.9
zipp==3.8.0
